Changes to be committed: 	new file:   ConTExt/SummarizeOutput.py 	modified:   ConTExt/tools/CoverageModeller.py 	modified:   ConTExt/tools/General_Tools.py 	modified:   ConTExt/tools/MapQuality.py
I repackaged the functions necessary for classifying the junctions identified in
in ExpMax as SummarizeOutput.py. I moved a function to CoverageModeller.py. I added an additional class to General_Tools which is needed by SummarizeOutput. I added a preliminary function to MapQuality to estimates the Mapping Quality of alignments to the consensus sequences; this appears
to be working, but it is somewhat slow as written. I think an additional approximation can speed it up.